Move notification iter test
Move the Notification iterator tests to the notification module.
